[refactor] move JSON.meta parsing code
There is no change in behaviour, but a bunch of TS errors as
the (untyped) functions are now in TS. We'll add types in the
next commit
